requirements.txt
livekit >= 0.17.1
livekit-agents~=0.10.0
livekit-plugins-openai~=0.10.0
python-dotenv~=1.0
